Revert "Revert "[SpellInfo]: use max targets per  difficult.""
This reverts commit 9173864fb87c7b1e964b2b197f0181dab52cec4b.

Conflicts:
	src/server/game/Spells/SpellMgr.cpp

+++ src/server/game/DataStores/DBCStores.cpp
@@ -1001,7 +1001,7 @@ SpellTotemsEntry const* GetSpellTotemEntry(uint32 spellId, uint8 totem)
     return itr->second.totems[totem];
 }
 
-SpellTargetRestrictionsEntry const *GetSpellTargetRestrioctions(uint32 spellId, uint8 difficulty)
+SpellTargetRestrictionsEntry const *GetSpellTargetRestrioctions(uint32 spellId, uint16 difficulty)
 {
     if(!difficulty)
         return NULL;
